third_party/WebKit/LayoutTests/http/tests/performance-timing/longtask-v2/longtask-executescript.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>LongTask Timing: validate the long task attributes for execute-script</title>
<body>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<h1>Long Task Attributes for Execute Script</h1>
<div id="log"></div>
<script>
  async_test(function (t) {
    const observer = new PerformanceObserver(
      t.step_func(function (entryList) {
        const entries = entryList.getEntries();
        assert_equals(entries.length, 1,
            'Exactly one entry is expected.');
        const longtask = entries[0];
        assert_equals(longtask.entryType, 'longtask');
        assert_equals(longtask.name, 'self');
        assert_greater_than(longtask.duration, 50);
        assert_greater_than_equal(longtask.startTime, 0);
        const currentTime = performance.now();
        assert_less_than_equal(longtask.startTime, currentTime);

        const it = longtask.attribution.find(function(it) {
          return it.name === 'script-run';
        });
        assert_false(!it, 'script-run is not captured.');
        assert_greater_than(it.duration, 51);
        assert_greater_than(it.startTime, 0);
        assert_equals(it.entryType, 'taskattribution');
        assert_equals(it.containerType, 'iframe');
        assert_equals(it.containerId, '');
        assert_equals(it.containerName, '');
        assert_equals(it.containerSrc, '');
        assert_true(it.scriptURL.includes('longtask-executescript.html'));
        observer.disconnect();
        t.done();
      })
    );
    observer.observe({entryTypes: ['longtask']});

    /* Generate a slow task. */
    const begin = window.performance.now();
    while (window.performance.now() < begin + 51);

}, 'Performance longtask entries with execute-script attribute are observable.');
</script>

</body>
